+++ Jarvis.Framework.Shared/HealthCheck/JarvisFrameworkHealthCheckResult.cs
@@ -8,7 +8,7 @@ namespace Jarvis.Framework.Shared.HealthCheck
     /// This is the very same class of Metrics.NET library, copied into framework to 
     /// make the transition from Metrics.NET to App.Metrics easier.
     /// </summary>
-    public struct HealthCheckResult
+    internal struct JarvisFrameworkHealthCheckResult
     {
         /// <summary>
         /// True if the check was successful, false if the check failed.
@@ -20,7 +20,7 @@ namespace Jarvis.Framework.Shared.HealthCheck
         /// </summary>
         public readonly string Message;
 
-        private HealthCheckResult(bool isHealthy, string message)
+        private JarvisFrameworkHealthCheckResult(bool isHealthy, string message)
         {
             IsHealthy = isHealthy;
             Message = message;
@@ -30,7 +30,7 @@ namespace Jarvis.Framework.Shared.HealthCheck
         /// Create a healthy status response.
         /// </summary>
         /// <returns>Healthy status response.</returns>
-        public static HealthCheckResult Healthy()
+        public static JarvisFrameworkHealthCheckResult Healthy()
         {
             return Healthy("OK");
         }
@@ -41,9 +41,9 @@ namespace Jarvis.Framework.Shared.HealthCheck
         /// <param name="message">Status message.</param>
         /// <param name="values">Values to format the status message with.</param>
         /// <returns>Healthy status response.</returns>
-        public static HealthCheckResult Healthy(string message)
+        public static JarvisFrameworkHealthCheckResult Healthy(string message)
         {
-            return new HealthCheckResult(true, string.IsNullOrWhiteSpace(message) ? "OK" : message);
+            return new JarvisFrameworkHealthCheckResult(true, string.IsNullOrWhiteSpace(message) ? "OK" : message);
         }
 
         /// <summary>
@@ -52,17 +52,17 @@ namespace Jarvis.Framework.Shared.HealthCheck
         /// <param name="message">Status message.</param>
         /// <param name="values">Values to format the status message with.</param>
         /// <returns>Healthy status response.</returns>
-        public static HealthCheckResult Healthy(string message, params object[] values)
+        public static JarvisFrameworkHealthCheckResult Healthy(string message, params object[] values)
         {
             var status = string.Format(message, values);
-            return new HealthCheckResult(true, string.IsNullOrWhiteSpace(status) ? "OK" : status);
+            return new JarvisFrameworkHealthCheckResult(true, string.IsNullOrWhiteSpace(status) ? "OK" : status);
         }
 
         /// <summary>
         /// Create a unhealthy status response.
         /// </summary>
         /// <returns>Unhealthy status response.</returns>
-        public static HealthCheckResult Unhealthy()
+        public static JarvisFrameworkHealthCheckResult Unhealthy()
         {
             return Unhealthy("FAILED");
         }
@@ -73,9 +73,9 @@ namespace Jarvis.Framework.Shared.HealthCheck
         /// <param name="message">Status message.</param>
         /// <param name="values">Values to format the status message with.</param>
         /// <returns>Unhealthy status response.</returns>
-        public static HealthCheckResult Unhealthy(string message)
+        public static JarvisFrameworkHealthCheckResult Unhealthy(string message)
         {
-            return new HealthCheckResult(false, string.IsNullOrWhiteSpace(message) ? "FAILED" : message);
+            return new JarvisFrameworkHealthCheckResult(false, string.IsNullOrWhiteSpace(message) ? "FAILED" : message);
         }
 
         /// <summary>
@@ -84,10 +84,10 @@ namespace Jarvis.Framework.Shared.HealthCheck
         /// <param name="message">Status message.</param>
         /// <param name="values">Values to format the status message with.</param>
         /// <returns>Unhealthy status response.</returns>
-        public static HealthCheckResult Unhealthy(string message, params object[] values)
+        public static JarvisFrameworkHealthCheckResult Unhealthy(string message, params object[] values)
         {
             var status = string.Format(message, values);
-            return new HealthCheckResult(false, string.IsNullOrWhiteSpace(status) ? "FAILED" : status);
+            return new JarvisFrameworkHealthCheckResult(false, string.IsNullOrWhiteSpace(status) ? "FAILED" : status);
         }
 
         /// <summary>
@@ -95,10 +95,10 @@ namespace Jarvis.Framework.Shared.HealthCheck
         /// </summary>
         /// <param name="exception">Exception to use for reason.</param>
         /// <returns>Unhealthy status response.</returns>
-        public static HealthCheckResult Unhealthy(Exception exception)
+        public static JarvisFrameworkHealthCheckResult Unhealthy(Exception exception)
         {
             var status = $"EXCEPTION: {exception.GetType().Name} - {exception.Message}";
-            return new HealthCheckResult(false, status + Environment.NewLine + FormatStackTrace(exception));
+            return new JarvisFrameworkHealthCheckResult(false, status + Environment.NewLine + FormatStackTrace(exception));
         }
 
         private static string FormatStackTrace(Exception exception, int indent = 2)