+++ Jarvis.Framework/Commands/AbstractCommandHandler.cs
@@ -1,15 +1,10 @@
 ﻿using Castle.Core.Logging;
-using Jarvis.Framework.Kernel.Commands;
-using Jarvis.Framework.Kernel.Engine;
-using Jarvis.Framework.Kernel.Events;
 using Jarvis.Framework.Shared;
 using Jarvis.Framework.Shared.Claims;
 using Jarvis.Framework.Shared.Commands;
 using Jarvis.Framework.Shared.Events;
-using Jarvis.Framework.Shared.IdentitySupport;
 using Jarvis.Framework.Shared.Logging;
 using Jarvis.Framework.Shared.Messages;
-using Jarvis.Framework.Shared.Persistence;
 using Jarvis.Framework.Shared.Support;
 using NStore.Domain;
 using System;
@@ -45,17 +40,11 @@ namespace Jarvis.Framework.Kernel.Commands
             using (var x = SecurityContextManager.SetCurrentClaims(claims))
             {
                 await OnCheckSecurityAsync(cmd).ConfigureAwait(false);
-                if (JarvisFrameworkGlobalConfiguration.KernelMetrics)
-                {
-                    using (var context = SharedMetricsHelper.StartCommandTimer(cmd))
-                    {
-                        await Execute(cmd).ConfigureAwait(false);
-                        SharedMetricsHelper.IncrementCommandCounter(cmd, context.Elapsed);
-                    }
-                }
-                else
+
+                using (var context = JarvisFrameworkSharedMetricsHelper.StartCommandTimer(cmd))
                 {
                     await Execute(cmd).ConfigureAwait(false);
+                    JarvisFrameworkSharedMetricsHelper.MarkCommandExecuted(cmd);
                 }
             }
         }