+++ Jarvis.Framework/ProjectionEngine/Rebuild/RebuildProjectionEngineV2.cs
@@ -95,16 +95,16 @@ namespace Jarvis.Framework.Kernel.ProjectionEngine.Rebuild
 
             _commitCollection = commitDatabase.GetCollection<MongoChunk>("Commits");
 
-            HealthChecks.RegisterHealthCheck("RebuildProjectionEngineV2", HealthCheck);
+            JarvisFrameworkHealthChecks.RegisterHealthCheck("RebuildProjectionEngineV2", HealthCheck);
         }
 
-        private HealthCheckResult HealthCheck()
+        private JarvisFrameworkHealthCheckResult HealthCheck()
         {
             if (String.IsNullOrEmpty(_pollError))
             {
-                return HealthCheckResult.Healthy();
+                return JarvisFrameworkHealthCheckResult.Healthy();
             }
-            return HealthCheckResult.Unhealthy(_pollError);
+            return JarvisFrameworkHealthCheckResult.Unhealthy(_pollError);
         }
 
         private void DumpProjections()
@@ -178,7 +178,7 @@ namespace Jarvis.Framework.Kernel.ProjectionEngine.Rebuild
 
                 //Create the dispatcher, a dispatcher is an object that is able to dispatch events to projections.
                 var dispatcher = new RebuildProjectionSlotDispatcher(Logger, slotName, _config, projectionsForThisSlot, maximumDispatchedValue);
-                KernelMetricsHelper.SetCheckpointCountToDispatch(slotName, () => dispatcher.CheckpointToDispatch);
+                JarvisFrameworkKernelMetricsHelper.SetCheckpointCountToDispatch(slotName, () => dispatcher.CheckpointToDispatch);
 
                 //For this slot we need to know the bucket that can dispatch it, the rule is simple, find the bucket that contains
                 //this exact slot name, and if you are unable to find that, find the bucket that has * (default bucket.)
@@ -243,7 +243,7 @@ namespace Jarvis.Framework.Kernel.ProjectionEngine.Rebuild
                             dispatcher.SlotName,
                             eventsOfThisSlot);
                     consumers.Add(slotInfo);
-                    KernelMetricsHelper.CreateMeterForRebuildDispatcherBuffer(dispatcher.SlotName, () => actionBlock.InputCount);
+                    JarvisFrameworkKernelMetricsHelper.CreateMeterForRebuildDispatcherBuffer(dispatcher.SlotName, () => actionBlock.InputCount);
                 }
                 var allTypeHandledStringList = _projectionInspector
                     .EventHandled
@@ -253,10 +253,10 @@ namespace Jarvis.Framework.Kernel.ProjectionEngine.Rebuild
                 var broadcaster = SlotGuaranteedDeliveryBroadcastBlock.CreateForStandardEvents(consumers, bucketInfo, consumer.Key.BufferSize);
                 _buffer.LinkTo(broadcaster, new DataflowLinkOptions() { PropagateCompletion = true });
 
-                KernelMetricsHelper.CreateGaugeForRebuildFirstBuffer(bucketInfo, () => _buffer.Count);
-                KernelMetricsHelper.CreateGaugeForRebuildBucketDBroadcasterBuffer(bucketInfo, () => broadcaster.InputCount);
+                JarvisFrameworkKernelMetricsHelper.CreateGaugeForRebuildFirstBuffer(bucketInfo, () => _buffer.Count);
+                JarvisFrameworkKernelMetricsHelper.CreateGaugeForRebuildBucketDBroadcasterBuffer(bucketInfo, () => broadcaster.InputCount);
 
-                KernelMetricsHelper.CreateMeterForRebuildEventCompleted(bucketInfo);
+                JarvisFrameworkKernelMetricsHelper.CreateMeterForRebuildEventCompleted(bucketInfo);
 
                 //fire each bucket in own thread
 #pragma warning disable CS4014 // Because this call is not awaited, execution of the current method continues before the call is completed
@@ -271,7 +271,7 @@ namespace Jarvis.Framework.Kernel.ProjectionEngine.Rebuild
 #pragma warning restore CS4014 // Because this call is not awaited, execution of the current method continues before the call is completed
             }
 
-            KernelMetricsHelper.SetProjectionEngineCurrentDispatchCount(() => RebuildProjectionMetrics.CountOfConcurrentDispatchingCommit);
+            JarvisFrameworkKernelMetricsHelper.SetProjectionEngineCurrentDispatchCount(() => RebuildProjectionMetrics.CountOfConcurrentDispatchingCommit);
             return _status;
         }